Share the sql_queries spec helper; tighten the preload assertion
Promote the sql_queries(matching:) capture helper out of plugin_routes_spec
into spec/support so query-shape specs reuse it. Rewrite the
verify_front_visibility non-promotion example to assert eager_loading? is
false (the property that matters -- a single-association includes/eager_load
promotion slips past an eager_load_values-only check), drop the hand-rolled
subscriber, and correct the comment: post_type.the_posts is PostType's plain
has_many (no join); the one tolerated LEFT OUTER JOIN is the categories
through-preload's term_taxonomy STI condition.
